версия 1.2.3
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@597 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/general/eioupdateindicator.sql
@@ -1,4 +1,4 @@
-create or replace function eioUpdateIndicator(int8, int4, varchar, timestamp, timestamp, int4, int4, varchar) returns int4 as
+create or replace function eioUpdateIndicator(int8, int4, varchar, timestamp, timestamp, int4, int4, varchar) returns int8 as
 $BODY$
 declare
     idRec alias for $1;
@@ -10,47 +10,47 @@ declare
     iIdObjectSrc1 alias for $7;
     iDesc alias for $8;
 
-    ok int4;
+    idRecAttrValue int8;
 begin
 
 
-    if(ioValue isnull) then
+    idRecAttrValue = eioCatIndExist(idRec, idCategoryAttr);
+    if(idRecAttrValue is not null and ioValue isnull) then
         delete from rec_attrs_values where id_record = idRec and id_attr_category = idCategoryAttr;
-        return 1;
+        return 0;
     end if;
 
-    if(eioCatIndExist(idRec, idCategoryAttr) = true) then
+    if(idRecAttrValue is not null) then
         update rec_attrs_values set 
             value = ioValue, 
             id_io_object_src = iIdObjectSrc, 
             id_io_object_src1 = iIdObjectSrc1, 
             description = iDesc 
         where 
-            id_record = idRec
-            and id_attr_category = idCategoryAttr;
+            id = idRecAttrValue;
     else
-        select eioInsertIndicator(idRec, idCategoryAttr, ioValue, iStartTime, iStopTime, iIdObjectSrc, iIdObjectSrc1, iDesc) into ok;
-        return ok;
+        select eioInsertIndicator(idRec, idCategoryAttr, ioValue, iStartTime, iStopTime, iIdObjectSrc, iIdObjectSrc1, iDesc) into idRecAttrValue;
+        return idRecAttrValue;
     end if;
 
-    return 1;
+    return idRecAttrValue;
 end
 $BODY$
 language 'plpgsql';
 
-create or replace function eioCatIndExist(int8, int4) returns bool as
+create or replace function eioCatIndExist(int8, int4) returns int8 as
 $BODY$
 declare
     idRec alias for $1;
     idCategoryAttr alias for $2;
-    cnt int4;
+    idRecAttrValue int8;
 begin
-    select count(*) into cnt from rec_attrs_values where id_record = idRec and id_attr_category = idCategoryAttr;
-    if(cnt > 0) then
-        return true;
+    select id into idRecAttrValue from rec_attrs_values where id_record = idRec and id_attr_category = idCategoryAttr limit 1;
+    if(idRecAttrValue isnull) then
+        return NULL;
     end if;
 
-    return false;
+    return idRecAttrValue;
 end
 $BODY$
 language 'plpgsql' security definer;
